Implement deletion of items. Currently only ids of deleted items are stored in an attribute of the collection. The purge method will come later.
svn path=/trunk/KDE/kdepim/akonadi/; revision=977471

+++ resources/mbox/libmbox/mbox.cpp
@@ -94,7 +94,7 @@ void MBox::close()
   d->mFileLocked = false;
 }
 
-QList<MsgInfo> MBox::entryList(const QSet<int> &deletedItems) const
+QList<MsgInfo> MBox::entryList(const QSet<quint64> &deletedItems) const
 {
   Q_ASSERT(d->mMboxFile.isOpen());
